+++ src/errors/reserve_error.rs
@@ -1,14 +1,14 @@
 use core::{error::Error, fmt};
 
-/// An error if allocating or resizing a [`LeanString`] failed.
+/// An error if allocating or resizing a [`CharString`] failed.
 ///
-/// [`LeanString`]: crate::LeanString
+/// [`CharString`]: crate::CharString
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct ReserveError;
 
 impl fmt::Display for ReserveError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("Cannot allocate memory to hold LeanString")
+        f.write_str("Cannot allocate memory to hold CharString")
     }
 }