+++ src/features/arbitrary.rs
@@ -1,14 +1,14 @@
-use crate::LeanString;
+use crate::CharString;
 use arbitrary::{Arbitrary, Result, Unstructured};
 
 #[cfg_attr(docsrs, doc(cfg(feature = "arbitrary")))]
-impl<'a> Arbitrary<'a> for LeanString {
+impl<'a> Arbitrary<'a> for CharString {
     fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
-        <&str as Arbitrary>::arbitrary(u).map(LeanString::from)
+        <&str as Arbitrary>::arbitrary(u).map(CharString::from)
     }
 
     fn arbitrary_take_rest(u: Unstructured<'a>) -> Result<Self> {
-        <&str as Arbitrary>::arbitrary_take_rest(u).map(LeanString::from)
+        <&str as Arbitrary>::arbitrary_take_rest(u).map(CharString::from)
     }
 
     fn size_hint(depth: usize) -> (usize, Option<usize>) {