+++ tests/property.rs
@@ -1,4 +1,4 @@
-use lean_string::{LeanString, ToLeanString};
+use char_str::{CharString, ToCharString};
 use proptest::{prelude::*, property_test};
 
 #[property_test]
@@ -6,14 +6,14 @@ use proptest::{prelude::*, property_test};
 fn create_from_str(input: String) {
     let str = input.as_str();
 
-    let lean = LeanString::from(str);
-    prop_assert_eq!(&lean, str);
-    prop_assert_eq!(lean.len(), str.len());
+    let actual = CharString::from(str);
+    prop_assert_eq!(&actual, str);
+    prop_assert_eq!(actual.len(), str.len());
 
     if str.len() <= 2 * size_of::<usize>() {
-        prop_assert!(!lean.is_heap_allocated());
+        prop_assert!(!actual.is_heap_allocated());
     } else {
-        prop_assert!(lean.is_heap_allocated());
+        prop_assert!(actual.is_heap_allocated());
     }
 }
 
@@ -22,16 +22,16 @@ fn create_from_str(input: String) {
 fn create_from_u8_bytes(input: Vec<u8>) {
     let bytes = input.as_slice();
 
-    let lean = LeanString::from_utf8(bytes);
+    let actual = CharString::from_utf8(bytes);
     let string = String::from_utf8(bytes.to_vec());
-    prop_assert_eq!(lean.is_err(), string.is_err());
-    if let (Ok(lean), Ok(string)) = (lean, string) {
-        prop_assert_eq!(&lean, &string);
+    prop_assert_eq!(actual.is_err(), string.is_err());
+    if let (Ok(actual), Ok(string)) = (actual, string) {
+        prop_assert_eq!(&actual, &string);
     }
 
-    let lean = LeanString::from_utf8_lossy(bytes);
+    let actual = CharString::from_utf8_lossy(bytes);
     let string = String::from_utf8_lossy(bytes);
-    prop_assert_eq!(&lean, &string);
+    prop_assert_eq!(&actual, &string);
 }
 
 #[property_test]
@@ -39,109 +39,109 @@ fn create_from_u8_bytes(input: Vec<u8>) {
 fn create_from_u16_bytes(input: Vec<u16>) {
     let bytes = input.as_slice();
 
-    let lean = LeanString::from_utf16(bytes);
+    let actual = CharString::from_utf16(bytes);
     let string = String::from_utf16(bytes);
-    prop_assert_eq!(lean.is_err(), string.is_err());
-    if let (Ok(lean), Ok(string)) = (lean, string) {
-        prop_assert_eq!(&lean, &string);
+    prop_assert_eq!(actual.is_err(), string.is_err());
+    if let (Ok(actual), Ok(string)) = (actual, string) {
+        prop_assert_eq!(&actual, &string);
     }
 
-    let lean = LeanString::from_utf16_lossy(bytes);
+    let actual = CharString::from_utf16_lossy(bytes);
     let string = String::from_utf16_lossy(bytes);
-    prop_assert_eq!(&lean, &string);
+    prop_assert_eq!(&actual, &string);
 }
 
 #[property_test]
 #[cfg_attr(miri, ignore)]
 fn collect_from_chars(input: String) {
-    let lean = input.chars().collect::<LeanString>();
-    prop_assert_eq!(&lean, &input);
+    let actual = input.chars().collect::<CharString>();
+    prop_assert_eq!(&actual, &input);
 }
 
 #[property_test]
 #[cfg_attr(miri, ignore)]
 fn collect_from_strings(input: Vec<String>) {
-    let lean = input.clone().into_iter().collect::<LeanString>();
+    let actual = input.clone().into_iter().collect::<CharString>();
     let string = input.into_iter().collect::<String>();
-    prop_assert_eq!(&lean, &string);
+    prop_assert_eq!(&actual, &string);
 }
 
-macro_rules! test_integer_to_lean_string {
+macro_rules! test_integer_to_char_string {
     ($($ty:ty),* $(,)?) => {$(
         paste::paste! {
             #[test]
-            fn [<$ty _to_lean_string>]() {
+            fn [<$ty _to_char_string>]() {
                 for num in <$ty>::MIN..=<$ty>::MAX {
-                    let lean = num.to_lean_string();
+                    let actual = num.to_char_string();
                     let string = num.to_string();
-                    assert_eq!(lean, string);
+                    assert_eq!(actual, string);
                 }
             }
             #[test]
-            fn [<nonzero_ $ty _to_lean_string>]() {
+            fn [<nonzero_ $ty _to_char_string>]() {
                 for num in <$ty>::MIN..=<$ty>::MAX {
                     if num == 0 { continue };
                     let num = core::num::NonZero::<$ty>::new(num).unwrap();
-                    let lean = num.to_lean_string();
+                    let actual = num.to_char_string();
                     let string = num.to_string();
-                    assert_eq!(lean, string);
+                    assert_eq!(actual, string);
                 }
             }
         }
     )*};
 }
-test_integer_to_lean_string!(u8, i8);
+test_integer_to_char_string!(u8, i8);
 
-macro_rules! prop_test_integer_to_lean_string {
+macro_rules! prop_test_integer_to_char_string {
     ($($ty:ty),* $(,)?) => {$(
         paste::paste! {
             #[property_test]
             #[cfg_attr(miri, ignore)]
-            fn [<$ty _to_lean_string>](i: $ty) {
-                prop_assert_eq!(i.to_lean_string(), i.to_string());
+            fn [<$ty _to_char_string>](i: $ty) {
+                prop_assert_eq!(i.to_char_string(), i.to_string());
             }
             #[property_test]
             #[cfg_attr(miri, ignore)]
-            fn [<nonzero_ $ty _to_lean_string>](i: core::num::NonZero<$ty>) {
-                prop_assert_eq!(i.to_lean_string(), i.to_string());
+            fn [<nonzero_ $ty _to_char_string>](i: core::num::NonZero<$ty>) {
+                prop_assert_eq!(i.to_char_string(), i.to_string());
             }
         }
     )*};
 }
-prop_test_integer_to_lean_string!(u16, i16, u32, i32, u64, i64, u128, i128, usize, isize);
+prop_test_integer_to_char_string!(u16, i16, u32, i32, u64, i64, u128, i128, usize, isize);
 
 #[property_test]
 #[cfg_attr(miri, ignore)]
-fn f32_to_lean_string(f: f32) {
-    let lean = f.to_lean_string();
-    let float = lean.parse::<f32>().unwrap();
+fn f32_to_char_string(f: f32) {
+    let actual = f.to_char_string();
+    let float = actual.parse::<f32>().unwrap();
     prop_assert_eq!(f, float);
 }
 
 #[property_test]
 #[cfg_attr(miri, ignore)]
-fn f64_to_lean_string(f: f64) {
-    let lean = f.to_lean_string();
-    let float = lean.parse::<f64>().unwrap();
+fn f64_to_char_string(f: f64) {
+    let actual = f.to_char_string();
+    let float = actual.parse::<f64>().unwrap();
     prop_assert_eq!(f, float);
 }
 
 #[test]
-fn bool_to_lean_string() {
+fn bool_to_char_string() {
     let t = true;
     let f = false;
-    assert_eq!(t.to_lean_string(), t.to_string());
-    assert_eq!(f.to_lean_string(), f.to_string());
+    assert_eq!(t.to_char_string(), t.to_string());
+    assert_eq!(f.to_char_string(), f.to_string());
 }
 
 #[property_test]
 #[cfg_attr(miri, ignore)]
-fn char_to_lean_string(c: char) {
-    prop_assert_eq!(c.to_lean_string(), c.to_string());
+fn char_to_char_string(c: char) {
+    prop_assert_eq!(c.to_char_string(), c.to_string());
 }
 
 #[property_test]
 #[cfg_attr(miri, ignore)]
-fn string_to_lean_string(s: String) {
-    prop_assert_eq!(s.to_lean_string(), s);
+fn string_to_char_string(s: String) {
+    prop_assert_eq!(s.to_char_string(), s);
 }